Add alternative Ropsten faucets. Link to KB
The more well-known ropsten.be faucet was glitchy recently
and not sending ropsten ether for days, so I had to look for other faucets
that work.
Added a KB article with two faucets so that people can get ropsten if one is down.
Linked all ropsten faucet mentions from the docs to the kb article to maintain
the list in one place.